Add Session tab for Ableton performance
