+++ frontend/src/components/layout/CenterTabBar.tsx
@@ -7,6 +7,7 @@ import {
   Workflow,
   Tv2,
   Disc,
+  Rows3,
 } from 'lucide-react';
 import { type CenterTab } from '../../state/appUiStore';
 
@@ -33,7 +34,7 @@ const TABS: Array<{
    *  so each workspace gets a recognizable color at a glance. */
   accent: { border: string; bg: string; text: string; iconText: string };
 }> = [
-  // Order locked by user: MAKE, EDIT, MIX, DJ, VJ, TRAIN, LEARN.
+  // Order locked by user: MAKE, EDIT, SESSION, MIX, DJ, VJ, TRAIN, LEARN.
   {
     id: 'make',
     label: 'Make',
@@ -51,6 +52,18 @@ const TABS: Array<{
     label: 'Edit',
     desc: 'Arrange clips on a timeline, add effects and automation, export',
     icon: Scissors,
+    accent: {
+      border: 'border-emerald-500/50',
+      bg: 'bg-emerald-500/15',
+      text: 'text-emerald-100',
+      iconText: 'text-emerald-300',
+    },
+  },
+  {
+    id: 'session',
+    label: 'Session',
+    desc: 'Ableton Session View import and scene performance',
+    icon: Rows3,
     accent: {
       border: 'border-emerald-500/50',
       bg: 'bg-emerald-500/15',
@@ -162,4 +175,3 @@ export const CenterTabBar: React.FC<CenterTabBarProps> = ({
     </div>
   );
 };
-
